refactor: 직책 수정에 따른 Role 수정 함수 위치 변경
